services/doc_mgmt_service/requirements.txt
python-dotenv
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-CORS
python-magic
werkzeug
